requirements.txt
-i https://pypi.org/simple
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
deprecated==1.2.13
dj-database-url==0.5.0
django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
flower==1.1.0
gunicorn==20.1.0
humanize==4.2.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.2
kombu==5.2.4
markupsafe==2.1.1
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pyjwt==2.4.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.1
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
setuptools==63.2.0
six==1.16.0
sqlparse==0.4.2
tornado==6.2
uritemplate==4.1.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
